% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtm_package.R
\docType{package}
\name{gtm}
\alias{gtm2}
\alias{gtm2-package}
\alias{gtm}
\alias{gtm_package}
\title{gtm: Modeling Group Testing Data}
\description{
The gtm package provides tools for modeling and analyzing group testing data.
It includes functions for fitting models, evaluating performance, and visualizing results.
}
\seealso{
\code{\link{dorfman}} for details on specific methods.
}
\author{
\strong{Maintainer}: Muhammad Yaseen \email{myaseen208@gmail.com} (\href{https://orcid.org/0000-0002-5923-1714}{ORCID}) (School of Mathematical & Statistical Sciences, Clemson University, Clemson, South Carolina, USA)

Authors:
\itemize{
  \item Christopher McMahan \email{mcmaha2@clemson.edu} (\href{https://orcid.org/0000-0001-5056-9615}{ORCID}) (School of Mathematical & Statistical Sciences, Clemson University, Clemson, South Carolina, USA)
  \item Christopher Bilder \email{bilder@unl.edu} (\href{https://orcid.org/0000-0002-2848-8576}{ORCID}) (Department of Statistics, University of Nebraska-Lincoln, Nebraska, USA)
  \item Joshua Tebbs \email{tebbs@stat.sc.edu} (\href{https://orcid.org/0000-0002-6762-7241}{ORCID}) (Department of Statistics, University of South Carolina, Columbia, South Carolina, USA)
}

}
\keyword{package}
